fix(dictation): ready-gate capture and unblock the fresh-tab mic
Live-test findings, traced to two root causes:

1. Mic dead on a new session tab: the composer hard-blocked dictation
   whenever the dictation.catalog probe was not exactly ready — a
   failed/racing probe latched the mic shut (also broke upstream
   browser-dictation e2e, which has no dictation.catalog fixture). The
   composer now only blocks on an explicit catalog unavailable;
   unknown/checking probes are inconclusive and let the press through.

2. Listening shown while the STT model was still loading: readiness
   is now gated on the server. The provider sends an open-time config
   message, readyOnOpen:false, and marks the session ready when the
   server replies `ready` (or first partial, or a 5s fallback for
   servers without the handshake). The relay forwards that as its
   existing `ready` event, and the composer now flips to listening
   only on true readiness instead of the create response.

Also: moved the status-render tests into chat-composer-status.test.ts
and shared test helpers into chat-composer.test-support.ts to stay
under lint file-size caps; typed the dictation onError params.
